Move _copy_file_or_tree tests from pyfunc utils to file utils
