+++ tests/pyfunc/test_model_export_with_loader_module_and_data_path.py
@@ -84,7 +84,7 @@ def test_model_log_load(sklearn_knn_model, iris_data, tmpdir):
         pyfunc_run_id = mlflow.active_run().info.run_uuid
 
     pyfunc_model_path = _get_model_log_dir(pyfunc_artifact_path, pyfunc_run_id)
-    model_config = Model.load(os.path.join(model_path, "MLmodel"))
+    model_config = Model.load(os.path.join(pyfunc_model_path, "MLmodel"))
     assert mlflow.pyfunc.FLAVOR_NAME in model_config.flavors
     assert mlflow.pyfunc.PY_VERSION in model_config.flavors[mlflow.pyfunc.FLAVOR_NAME]
     reloaded_model = mlflow.pyfunc.load_pyfunc(pyfunc_model_path)
